guide cue file structure
