+++ .changeset/curvy-windows-dance.md
@@ -1,0 +1,5 @@
+---
+'@leafygreen-ui/guide-cue': patch
+---
+
+Restructures package internals
