+++ packages/guide-cue/src/GuideCue/GuideCue.styles.ts
@@ -1,21 +1,12 @@
 import { transparentize } from 'polished';
 
 import { css, cx } from '@leafygreen-ui/emotion';
-import { Theme } from '@leafygreen-ui/lib';
 import { palette } from '@leafygreen-ui/palette';
 
 export const timeout1 = 400;
 export const timeout2 = 100;
 
 const size = 24;
-
-export const tooltipMultistepStyles = css`
-  padding: 32px 16px 16px;
-`;
-
-export const tooltipStyles = css`
-  cursor: auto;
-`;
 
 export const beaconStyles = (
   prefersReducedMotion: boolean,
@@ -123,56 +114,3 @@ export const beaconStyles = (
     `,
   );
 };
-
-export const contentStyles = css`
-  margin-bottom: 16px;
-`;
-
-export const footerStyles = css`
-  display: flex;
-  justify-content: flex-end;
-  align-items: center;
-  gap: 16px;
-`;
-
-export const bodyThemeStyles: Record<Theme, string> = {
-  [Theme.Light]: css`
-    color: ${palette.gray.light1};
-  `,
-  [Theme.Dark]: css`
-    color: ${palette.black};
-  `,
-};
-
-export const bodyTitleStyles = css`
-  margin-bottom: 4px;
-`;
-
-export const closeStyles = css`
-  position: absolute;
-  top: 10px;
-  right: 10px;
-`;
-
-export const closeHoverStyles = css`
-  color: ${palette.gray.dark2};
-  &:hover,
-  &:active {
-    &::before {
-      background-color: ${palette.gray.light3};
-    }
-  }
-`;
-
-export const stepStyles: Record<Theme, string> = {
-  [Theme.Light]: css`
-    color: ${palette.gray.base};
-  `,
-  [Theme.Dark]: css`
-    color: ${palette.gray.dark2};
-  `,
-};
-
-export const buttonStyles = css`
-  height: 28px;
-`;
