+++ packages/guide-cue/src/GuideCue/index.ts
@@ -1,0 +1,2 @@
+export { default } from './GuideCue';
+export { GuideCueProps, TooltipAlign, TooltipJustify } from './GuideCue.types';
