+++ packages/guide-cue/src/TooltipContent/index.ts
@@ -1,0 +1,1 @@
+export { default } from './TooltipContent';
